Extraction of methods for serializing BulkImport to separate classes
