+++ core/src/main/java/io/lionweb/lioncore/java/serialization/FlatBuffersSerialization.java
@@ -151,168 +151,6 @@ public class FlatBuffersSerialization extends AbstractSerialization {
         allNodes.stream().filter(n -> !(n instanceof ProxyNode)).collect(Collectors.toList()));
   }
 
-  public byte[] serializeBulkImport(BulkImport bulkImport) {
-    FlatBufferBuilder builder = new FlatBufferBuilder(1024);
-
-    FBHelper helper = new FBHelper(builder);
-    Map<String, String> containerByAttached = new HashMap<>();
-
-    int[] attachPointOffsets = new int[bulkImport.getAttachPoints().size()];
-    int i = 0;
-    for (BulkImport.AttachPoint attachPoint : bulkImport.getAttachPoints()) {
-      int containment = helper.offsetForMetaPointer(attachPoint.containment);
-      int container = builder.createSharedString(attachPoint.container);
-      int root = builder.createSharedString(attachPoint.rootId);
-      FBAttachPoint.startFBAttachPoint(builder);
-      FBAttachPoint.addContainer(builder, container);
-      FBAttachPoint.addContainment(builder, containment);
-      FBAttachPoint.addRoot(builder, root);
-      attachPointOffsets[i] = FBAttachPoint.endFBAttachPoint(builder);
-      i++;
-      containerByAttached.put(attachPoint.rootId, attachPoint.container);
-    }
-
-    int[] nodesOffsets = new int[bulkImport.getNodes().size()];
-    i = 0;
-    for (ClassifierInstance<?> node : bulkImport.getNodes()) {
-
-      List<Feature<?>> features = node.getClassifier().allFeatures();
-      List<Property> properties = new LinkedList<>();
-      List<Containment> containments = new LinkedList<>();
-      List<Reference> references = new LinkedList<>();
-      for (Feature<?> feature : features) {
-        if (feature instanceof Property) {
-          properties.add((Property) feature);
-        } else if (feature instanceof Containment) {
-          containments.add((Containment) feature);
-        } else if (feature instanceof Reference) {
-          references.add((Reference) feature);
-        } else {
-          throw new IllegalStateException();
-        }
-      }
-      int[] propOffsets;
-      {
-        propOffsets = new int[properties.size()];
-        int fi = 0;
-        for (Property property : properties) {
-          int metaPointer = helper.offsetForMetaPointer(MetaPointer.from(property));
-          Object propertyValue = node.getPropertyValue(property);
-          int propertyValueOffset = -1;
-          if (propertyValue != null) {
-            String propertyValueStr =
-                this.primitiveValuesSerialization.serialize(
-                    property.getType().getID(), propertyValue);
-            propertyValueOffset = builder.createSharedString(propertyValueStr);
-          }
-          FBProperty.startFBProperty(builder);
-          FBProperty.addMetaPointer(builder, metaPointer);
-          if (propertyValueOffset != -1) {
-            FBProperty.addValue(builder, propertyValueOffset);
-          }
-          propOffsets[fi] = FBProperty.endFBProperty(builder);
-          fi++;
-        }
-      }
-      int[] contOffsets;
-      {
-        contOffsets = new int[containments.size()];
-        int fi = 0;
-        for (Containment containment : containments) {
-          List<? extends Node> children = node.getChildren(containment);
-          int[] childOffsets = new int[children.size()];
-          int ci = 0;
-          for (Node child : children) {
-            childOffsets[ci] = builder.createSharedString(child.getID());
-            ci++;
-          }
-
-          int metaPointer = helper.offsetForMetaPointer(MetaPointer.from(containment));
-          int childrenVector = FBContainment.createChildrenVector(builder, childOffsets);
-          FBContainment.startFBContainment(builder);
-          FBContainment.addMetaPointer(builder, metaPointer);
-          FBContainment.addChildren(builder, childrenVector);
-          contOffsets[fi] = FBContainment.endFBContainment(builder);
-          fi++;
-        }
-      }
-      int[] refeOffsets;
-      {
-        refeOffsets = new int[references.size()];
-        int fi = 0;
-        for (Reference reference : references) {
-          List<ReferenceValue> referenceValues = node.getReferenceValues(reference);
-          int[] refValuesOffsets = new int[referenceValues.size()];
-          int ci = 0;
-          for (ReferenceValue referenceValue : referenceValues) {
-            int resolveInfo = -1;
-            if (referenceValue.getResolveInfo() != null) {
-              resolveInfo = builder.createSharedString(referenceValue.getResolveInfo());
-            }
-            int referredID = -1;
-            if (referenceValue.getReferredID() != null) {
-              referredID = builder.createSharedString(referenceValue.getReferredID());
-            }
-            FBReferenceValue.startFBReferenceValue(builder);
-            if (resolveInfo != -1) {
-              FBReferenceValue.addResolveInfo(builder, resolveInfo);
-            }
-            if (referredID != -1) {
-              FBReferenceValue.addReferred(builder, referredID);
-            }
-            refValuesOffsets[ci] = FBReferenceValue.endFBReferenceValue(builder);
-            ci++;
-          }
-
-          int metaPointer = helper.offsetForMetaPointer(MetaPointer.from(reference));
-          int valuesVector = FBReference.createValuesVector(builder, refValuesOffsets);
-          FBReference.startFBReference(builder);
-          FBReference.addMetaPointer(builder, metaPointer);
-          FBReference.addValues(builder, valuesVector);
-          refeOffsets[fi] = FBReference.endFBReference(builder);
-          fi++;
-        }
-      }
-      int[] annOffsets = new int[node.getAnnotations().size()];
-      int ai = 0;
-      for (AnnotationInstance annotation : node.getAnnotations()) {
-        annOffsets[ai] = builder.createSharedString(annotation.getID());
-        ai++;
-      }
-
-      int classifier = helper.offsetForMetaPointer(MetaPointer.from(node.getClassifier()));
-      int id = builder.createSharedString(node.getID());
-      int propsVector = FBNode.createPropertiesVector(builder, propOffsets);
-      int consVector = FBNode.createContainmentsVector(builder, contOffsets);
-      int refsVector = FBNode.createReferencesVector(builder, refeOffsets);
-      int annsVector = FBNode.createAnnotationsVector(builder, annOffsets);
-      String parentID =
-          node.getParent() == null
-              ? containerByAttached.get(node.getID())
-              : node.getParent().getID();
-      int parent = builder.createSharedString(parentID);
-      FBNode.startFBNode(builder);
-      FBNode.addId(builder, id);
-      FBNode.addClassifier(builder, classifier);
-      FBNode.addProperties(builder, propsVector);
-      FBNode.addContainments(builder, consVector);
-      FBNode.addReferences(builder, refsVector);
-      FBNode.addAnnotations(builder, annsVector);
-      FBNode.addParent(builder, parent);
-
-      nodesOffsets[i] = FBNode.endFBNode(builder);
-      i++;
-    }
-    int attachPointsVectorOffset =
-        FBBulkImport.createAttachPointsVector(builder, attachPointOffsets);
-    int nodesVectorOffset = FBBulkImport.createNodesVector(builder, nodesOffsets);
-    FBBulkImport.startFBBulkImport(builder);
-    FBBulkImport.addAttachPoints(builder, attachPointsVectorOffset);
-    FBBulkImport.addNodes(builder, nodesVectorOffset);
-    builder.finish(FBBulkImport.endFBBulkImport(builder));
-    return builder.dataBuffer().compact().array();
-  }
-
   public byte[] serializeNodesToByteArray(List<ClassifierInstance<?>> classifierInstances) {
     if (classifierInstances.stream().anyMatch(n -> n instanceof ProxyNode)) {
       throw new IllegalArgumentException("Proxy nodes cannot be serialized");
@@ -340,18 +178,18 @@ public class FlatBuffersSerialization extends AbstractSerialization {
     return serialize(serializedChunk);
   }
 
-  private class FBHelper {
+  protected class FBHelper {
     FlatBufferBuilder builder;
     Map<MetaPointer, Integer> serializedMetapointers = new HashMap<>();
 
-    FBHelper(FlatBufferBuilder builder) {
+    public FBHelper(FlatBufferBuilder builder) {
       this.builder = builder;
     }
 
     /**
      * This method can create objects, so it should not be nested inside another object creation.
      */
-    int offsetForMetaPointer(MetaPointer metaPointer) {
+    public int offsetForMetaPointer(MetaPointer metaPointer) {
       if (!serializedMetapointers.containsKey(metaPointer)) {
         int fbMetapointer =
             FBMetaPointer.createFBMetaPointer(
@@ -364,7 +202,7 @@ public class FlatBuffersSerialization extends AbstractSerialization {
       return serializedMetapointers.get(metaPointer);
     }
 
-    int[] languagesVector(List<UsedLanguage> usedLanguages) {
+    public int[] languagesVector(List<UsedLanguage> usedLanguages) {
       int[] languagesOffsets = new int[usedLanguages.size()];
       for (int i = 0; i < usedLanguages.size(); i++) {
         UsedLanguage ul = usedLanguages.get(i);
@@ -377,7 +215,7 @@ public class FlatBuffersSerialization extends AbstractSerialization {
       return languagesOffsets;
     }
 
-    int[] propsVector(List<SerializedPropertyValue> properties) {
+    public int[] propsVector(List<SerializedPropertyValue> properties) {
       int[] props = new int[properties.size()];
       for (int j = 0; j < properties.size(); j++) {
         SerializedPropertyValue el = properties.get(j);
@@ -390,7 +228,7 @@ public class FlatBuffersSerialization extends AbstractSerialization {
       return props;
     }
 
-    int[] containmentsVector(List<SerializedContainmentValue> containments) {
+    public int[] containmentsVector(List<SerializedContainmentValue> containments) {
       int[] cons = new int[containments.size()];
       for (int j = 0; j < containments.size(); j++) {
         SerializedContainmentValue el = containments.get(j);
@@ -411,7 +249,7 @@ public class FlatBuffersSerialization extends AbstractSerialization {
       return cons;
     }
 
-    int[] referencesVector(List<SerializedReferenceValue> references) {
+    public int[] referencesVector(List<SerializedReferenceValue> references) {
       int[] refs = new int[references.size()];
       for (int j = 0; j < references.size(); j++) {
         SerializedReferenceValue el = references.get(j);
@@ -432,7 +270,7 @@ public class FlatBuffersSerialization extends AbstractSerialization {
       return refs;
     }
 
-    int[] annotationsVector(List<String> annotations) {
+    public int[] annotationsVector(List<String> annotations) {
       int[] anns = new int[annotations.size()];
       if (anns.length > 0) {
         throw new UnsupportedOperationException();
