+++ core/src/main/java/io/lionweb/lioncore/java/serialization/ProtoBufSerialization.java
@@ -160,11 +160,13 @@ public class ProtoBufSerialization extends AbstractSerialization {
     return serialize(serializedChunk).toByteArray();
   }
 
-  private class SerializeHelper {
-    final Map<MetaPointer, Integer> metaPointers = new HashMap<>();
-    final Map<String, Integer> strings = new HashMap<>();
+  protected class SerializeHelper {
+    public final Map<MetaPointer, Integer> metaPointers = new HashMap<>();
+    public final Map<String, Integer> strings = new HashMap<>();
 
-    int stringIndexer(String string) {
+    public SerializeHelper() {}
+
+    public int stringIndexer(String string) {
       if (string == null) {
         return -1;
       }
@@ -176,7 +178,7 @@ public class ProtoBufSerialization extends AbstractSerialization {
       return index;
     };
 
-    int metaPointerIndexer(MetaPointer metaPointer) {
+    public int metaPointerIndexer(MetaPointer metaPointer) {
       if (metaPointers.containsKey(metaPointer)) {
         return metaPointers.get(metaPointer);
       }
@@ -191,7 +193,7 @@ public class ProtoBufSerialization extends AbstractSerialization {
       return index;
     }
 
-    PBNode serializeNode(SerializedClassifierInstance n) {
+    public PBNode serializeNode(SerializedClassifierInstance n) {
       PBNode.Builder nodeBuilder = PBNode.newBuilder();
       nodeBuilder.setId(this.stringIndexer(n.getID()));
       nodeBuilder.setClassifier(this.metaPointerIndexer((n.getClassifier())));
@@ -235,49 +237,6 @@ public class ProtoBufSerialization extends AbstractSerialization {
       n.getAnnotations().forEach(a -> nodeBuilder.addAnnotations(this.stringIndexer(a)));
       return nodeBuilder.build();
     }
-  }
-
-  public PBBulkImport serializeBulkImport(BulkImport bulkImport) {
-    PBBulkImport.Builder bulkImportBuilder = PBBulkImport.newBuilder();
-    ProtoBufSerialization.SerializeHelper serializeHelper =
-        new ProtoBufSerialization.SerializeHelper();
-
-    bulkImport
-        .getAttachPoints()
-        .forEach(
-            attachPoint -> {
-              PBAttachPoint.Builder attachPointBuilder = PBAttachPoint.newBuilder();
-              attachPointBuilder.setContainer(serializeHelper.stringIndexer(attachPoint.container));
-              attachPointBuilder.setRootId(serializeHelper.stringIndexer(attachPoint.rootId));
-              attachPointBuilder.setMetaPointerIndex(
-                  serializeHelper.metaPointerIndexer(attachPoint.containment));
-              bulkImportBuilder.addAttachPoints(attachPointBuilder.build());
-            });
-
-    SerializedChunk serializedChunk = serializeNodesToSerializationBlock(bulkImport.getNodes());
-
-    serializedChunk
-        .getClassifierInstances()
-        .forEach(
-            serializedNode -> {
-              bulkImportBuilder.addNodes(serializeHelper.serializeNode(serializedNode));
-            });
-
-    serializeHelper.strings.entrySet().stream()
-        .sorted(Map.Entry.comparingByValue())
-        .forEach(entry -> bulkImportBuilder.addStringValues(entry.getKey()));
-
-    serializeHelper.metaPointers.entrySet().stream()
-        .sorted(Map.Entry.comparingByValue())
-        .forEach(
-            entry ->
-                bulkImportBuilder.addMetaPointers(
-                    PBMetaPointer.newBuilder()
-                        .setLanguage(serializeHelper.stringIndexer(entry.getKey().getLanguage()))
-                        .setKey(serializeHelper.stringIndexer(entry.getKey().getKey()))
-                        .setVersion(serializeHelper.stringIndexer(entry.getKey().getVersion()))
-                        .build()));
-    return bulkImportBuilder.build();
   }
 
   public PBChunk serializeTree(ClassifierInstance<?> classifierInstance) {
