+++ core/src/main/java/io/lionweb/lioncore/java/serialization/SerializationProvider.java
@@ -2,6 +2,8 @@ package io.lionweb.lioncore.java.serialization;
 
 import io.lionweb.lioncore.java.language.LionCoreBuiltins;
 import io.lionweb.lioncore.java.self.LionCore;
+import io.lionweb.lioncore.java.serialization.extras.ExtraFlatBuffersSerialization;
+import io.lionweb.lioncore.java.serialization.extras.ExtraProtoBufSerialization;
 
 public class SerializationProvider {
 
@@ -32,6 +34,18 @@ public class SerializationProvider {
   /** This has specific support for LionCore or LionCoreBuiltins. */
   public static FlatBuffersSerialization getStandardFlatBuffersSerialization() {
     FlatBuffersSerialization serialization = new FlatBuffersSerialization();
+    standardInitialization(serialization);
+    return serialization;
+  }
+
+  public static ExtraFlatBuffersSerialization getExtraStandardFlatBuffersSerialization() {
+    ExtraFlatBuffersSerialization serialization = new ExtraFlatBuffersSerialization();
+    standardInitialization(serialization);
+    return serialization;
+  }
+
+  public static ExtraProtoBufSerialization getExtraStandardProtoBufSerialization() {
+    ExtraProtoBufSerialization serialization = new ExtraProtoBufSerialization();
     standardInitialization(serialization);
     return serialization;
   }
