+++ core/src/test/java/io/lionweb/lioncore/java/serialization/FlatbuffersSerializationTest.java
@@ -10,10 +10,6 @@ import io.lionweb.lioncore.java.model.Node;
 import io.lionweb.lioncore.java.model.impl.DynamicAnnotationInstance;
 import io.lionweb.lioncore.java.model.impl.DynamicNode;
 import io.lionweb.lioncore.java.serialization.data.*;
-import io.lionweb.lioncore.java.serialization.flatbuffers.FBAttachPoint;
-import io.lionweb.lioncore.java.serialization.flatbuffers.FBBulkImport;
-import io.lionweb.lioncore.java.serialization.flatbuffers.FBMetaPointer;
-import io.lionweb.lioncore.java.serialization.flatbuffers.FBNode;
 import io.lionweb.lioncore.java.serialization.refsmm.ContainerNode;
 import io.lionweb.lioncore.java.serialization.refsmm.RefNode;
 import io.lionweb.lioncore.java.serialization.refsmm.RefsLanguage;
@@ -21,7 +17,6 @@ import io.lionweb.lioncore.java.serialization.simplemath.IntLiteral;
 import io.lionweb.lioncore.java.serialization.simplemath.SimpleMathLanguage;
 import io.lionweb.lioncore.java.serialization.simplemath.Sum;
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -334,46 +329,6 @@ public class FlatbuffersSerializationTest extends SerializationTest {
     assertEquals(2, serializedChunk.getLanguages().size());
     assertSerializedChunkContainsLanguage(serializedChunk, l);
     assertSerializedChunkContainsLanguage(serializedChunk, LionCoreBuiltins.getInstance());
-  }
-
-  @Test
-  public void bulkImportSerialization() {
-    Language l = new Language("l", "l-id", "l-key", "1");
-    Concept c = new Concept(l, "c", "c-id", "c-key");
-    Property property = Property.createRequired("foo", LionCoreBuiltins.getString());
-    property.setID("p-id");
-    property.setKey("p-key");
-    c.addFeature(property);
-
-    DynamicNode n1 = new DynamicNode("n1", c);
-    ClassifierInstanceUtils.setPropertyValueByName(n1, "foo", "abc");
-
-    BulkImport bulkImport = new BulkImport();
-    bulkImport.addNode(n1);
-    bulkImport.addAttachPoint(
-        new BulkImport.AttachPoint("n2", new MetaPointer("Foo", "1", "c-key"), "n1"));
-
-    FlatBuffersSerialization flatBuffersSerialization =
-        SerializationProvider.getStandardFlatBuffersSerialization();
-    byte[] bytes = flatBuffersSerialization.serializeBulkImport(bulkImport);
-
-    ByteBuffer bb = ByteBuffer.wrap(bytes);
-    FBBulkImport fbBulkImport = FBBulkImport.getRootAsFBBulkImport(bb);
-    assertEquals(1, fbBulkImport.attachPointsLength());
-    assertEquals(1, fbBulkImport.nodesLength());
-
-    FBAttachPoint fbAttachPoint = fbBulkImport.attachPoints(0);
-
-    assertEquals("n2", fbAttachPoint.container());
-    assertEquals("n1", fbAttachPoint.root());
-
-    FBMetaPointer fbAttachPointContainment = fbAttachPoint.containment();
-    assertEquals("Foo", fbAttachPointContainment.language());
-    assertEquals("1", fbAttachPointContainment.version());
-    assertEquals("c-key", fbAttachPointContainment.key());
-
-    FBNode fbNode = fbBulkImport.nodes(0);
-    assertEquals("n1", fbNode.id());
   }
 
   private void assertSerializedChunkContainsLanguage(
