+++ core/src/test/java/io/lionweb/lioncore/java/serialization/SerializationTest.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 /** Base class with some utility methods used by several tests. */
-abstract class SerializationTest {
+public abstract class SerializationTest {
 
   protected List<JsonObject> getNodesByConcept(JsonArray nodes, String conceptKey) {
     return nodes.asList().stream()
